move abi util fn to solidity-abi crate
