+++ crates/util/solidity-abi/Cargo.toml
@@ -3,9 +3,9 @@ name = "solidity-abi"
 version = "0.1.0"
 edition = "2021"
 
-
 [dependencies]
 cfx-types = { workspace = true }
+cfx-bytes = { workspace = true }
 keccak-hash = { workspace = true }
 solidity-abi-derive = { workspace = true }
