+++ crates/util/solidity-abi/src/lib.rs
@@ -11,7 +11,9 @@ mod string;
 mod tests;
 mod utils;
 
-pub use self::utils::{read_abi_list, ABIListWriter, LinkedBytes};
+pub use self::utils::{
+    read_abi_list, string_revert_reason_decode, ABIListWriter, LinkedBytes,
+};
 use cfx_types::H256;
 use keccak_hash::keccak;
