+++ crates/util/solidity-abi/src/utils.rs
@@ -2,7 +2,8 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use super::{ABIDecodeError, ABIVariable};
+use super::{ABIDecodable, ABIDecodeError, ABIVariable};
+use cfx_bytes::Bytes;
 use cfx_types::U256;
 use std::{collections::LinkedList, slice::Iter};
 
@@ -127,4 +128,47 @@ pub fn pull_slice<'a>(
     let result = &slice[0..n];
     *iter = slice[n..].iter();
     Ok(result)
+}
+
+// abi decode string revert reason: Error(string)
+pub fn string_revert_reason_decode(output: &Bytes) -> String {
+    const MAX_LENGTH: usize = 50;
+    let decode_result = if output.len() < 4 {
+        Err(ABIDecodeError("Uncompleted Signature"))
+    } else {
+        let (sig, data) = output.split_at(4);
+        if sig != [8, 195, 121, 160] {
+            Err(ABIDecodeError("Unrecognized Signature"))
+        } else {
+            String::abi_decode(data)
+        }
+    };
+    match decode_result {
+        Ok(str) => {
+            if str.len() < MAX_LENGTH {
+                str
+            } else {
+                format!("{}...", str[..MAX_LENGTH].to_string())
+            }
+        }
+        Err(_) => "".to_string(),
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::string_revert_reason_decode;
+    use rustc_hex::FromHex;
+
+    #[test]
+    fn test_decode_result() {
+        let input_hex = "08c379a0\
+            0000000000000000000000000000000000000000000000000000000000000020\
+            0000000000000000000000000000000000000000000000000000000000000018\
+            e699bae59586e4b88de8b6b3efbc8ce8afb7e58585e580bc0000000000000000";
+        assert_eq!(
+            "智商不足，请充值".to_string(),
+            string_revert_reason_decode(&input_hex.from_hex().unwrap())
+        );
+    }
 }
